finalized the renaming of files, components, folders and imports
