+++ apps/lumina-core/src/app/editor/[[...slug]]/page.tsx
@@ -1,9 +1,7 @@
-import { Editor } from "@/components/editor/editor";
+import { LuminaEditor } from "@/components/lumina-editor/lumina-editor";
 import { getFullData } from "@/lib/dataFetcher";
-import { Render } from "@/components/render/render";
+import { LuminaRender } from "@/components/lumina-render/lumina-render";
 import { ContextProvider } from "@/context/contextProvider";
-import { TreeViewTab } from "@/components/tabs/treeView/treeView";
-import { PagesTab } from "@/components/lumina-tabs/pages/page";
 
 type Props = {
   params: {
@@ -26,9 +24,9 @@ export default async function EditorPage({ params }: Props) {
         },
       }}
     >
-      <Editor>
-        <Render />
-      </Editor>
+      <LuminaEditor>
+        <LuminaRender />
+      </LuminaEditor>
     </ContextProvider>
   );
 }
