+++ apps/lumina-core/src/components/linkBox/linkBox.tsx
@@ -3,8 +3,7 @@
 // does not support server side render
 
 import styles from '@/components/linkBox/linkBox.module.css'
-import { ShowEdit } from '@/components/showEdit/showEdit'
-import { useEffect, useState } from 'react'
+import { LuminaShowEdit } from '@/components/lumina-showEdit/lumina-showEdit'
 import { config } from './config'
 
 type TProps = {
@@ -25,7 +24,7 @@ export const LinkBox = ({id, title, description, href, color = 'black'}: TProps)
       target="_blank"
       rel="noopener noreferrer"
     >
-      <ShowEdit id={id} config={config} data={{ title, description, href, color}}/>
+      <LuminaShowEdit id={id} config={config} data={{ title, description, href, color}}/>
       <h2 style={{color}}>
         {title} <span>-&gt;</span>
       </h2>
