+++ apps/lumina-core/src/components/lumina-action-buttons/edit/lumina-edit-component.tsx
@@ -2,7 +2,7 @@ import { LuminaButton } from "@/components/lumina-button/lumina-button";
 import { LuminaEditModal } from "@/components/lumina-modals/edit/lumina-edit-modal";
 import { useLuminaContext } from "@/context/contextProvider";
 import { IComponentProps } from "@/data/data";
-import { TConfig } from "@/models/showEditModel";
+import { TConfig } from "@/lumina-models/lumina-showEditModel";
 import { useCallback, useState } from "react";
 
 type TProps = {
@@ -12,7 +12,7 @@ type TProps = {
   config: TConfig
 }
 
-export const EditComponentButton = ({ id, data, onUpdate, config }: TProps) => {
+export const LuminaEditComponentButton = ({ id, data, onUpdate, config }: TProps) => {
   const { dispatch } = useLuminaContext()
   const [showModalEdit, setShowModalEdit] = useState(false);
   const [formData, setFormData] = useState(data || {});
