+++ apps/lumina-core/src/components/lumina-action-buttons/visible/lumina-visible-component.tsx
@@ -1,6 +1,6 @@
 import { LuminaButton } from "@/components/lumina-button/lumina-button"
 
-export const VisibleComponentButton = () => {
+export const LuminaVisibleComponentButton = () => {
 
   return (
     <LuminaButton
