+++ apps/lumina-core/src/components/lumina-render/lumina-dynamicComponent.tsx
@@ -1,0 +1,6 @@
+import staticComponentsPath from "@/staticComponentsPath"
+
+export const LuminaDynamicComponent = (type: string) => {
+  if (!staticComponentsPath[type]) return null
+  return staticComponentsPath[type]
+}
