+++ apps/lumina-core/src/components/lumina-showEdit/readme.md
