+++ apps/lumina-core/src/components/lumina-tabs/treeView/lumina-treeView.tsx
@@ -1,83 +1,13 @@
 "use client";
 
 import styles from "./lumina-treeView.module.scss";
-import cx from "classnames";
 import { useLuminaContext } from "@/context/contextProvider";
-import { IComponentData, IComponentProps } from "@/data/data";
-import { useCallback, useState } from "react";
-import { ShowEdit } from "../../showEdit/showEdit";
 import { LuminaSearchBar } from "../../lumina-search/lumina-search";
-import { TreeviewHeader } from "./treeviewHeader/lumina-treeViewHeader";
-import { AddComponentButton } from "@/components/lumina-action-buttons/add/lumina-add-component";
+import { LuminaTreeviewHeader } from "./treeviewHeader/lumina-treeViewHeader";
+import { LuminaAddComponentButton } from "@/components/lumina-action-buttons/add/lumina-add-component";
+import { LuminaComponentTree } from "./lumina-componentTree";
 
-const TreeBranch = ({ data }: { data: IComponentData, noUp: boolean, noDown: boolean }) => {
-  const [showChildren, setShowChildren] = useState(false);
-
-  const handleTreeHeadClick = useCallback(() => {
-    setShowChildren(!showChildren);
-  }, [showChildren]);
-
-
-  const iconChange = () => {
-    if (data.children?.length) {
-      return (
-        <span className={cx(styles.treeViewIcon, showChildren ? 'lumina-close-up' : 'lumina-open-down', styles.treeViewPointer)} onClick={handleTreeHeadClick}></span>)
-    }
-    else {
-      return (
-        <span className={cx(styles.treeViewIcon, 'lumina-component')}></span>)
-    }
-  }
-
-  return (
-    <div className={styles.treeContainer}>
-      {iconChange()}
-      <div
-        className={cx(styles.treeHeadItem, {
-          [String(styles.pointerTreeView)]: data.children?.length,
-        })}
-      >
-        {data.type} - {data.friendlyName || data.id}{" "}
-        <ShowEdit
-          id={data.id}
-          inline={true}
-          data={data.props as IComponentProps}
-          visible={false}
-          noUp={false}
-          noDown={false}
-          menu={false}
-        />
-      </div>
-      {(data.children?.length && showChildren && (
-        <div className={styles.treeChildren}>
-          <ComponentTree data={data.children} />
-        </div>
-      ))}
-    </div>
-  );
-};
-
-const ComponentTree = ({ data }: { data: IComponentData[] }) => {
-  if (!data) return null;
-  return (
-    <>
-      {data.sort(
-        (a, b) => {
-          if (a.order < b.order) {
-            return -1;
-          } else if (a.order > b.order) {
-            return 1;
-          }
-          return 0;
-        }
-      ).map((dataItem, index) => (
-        <TreeBranch key={dataItem.id} data={dataItem} noUp={index === 0} noDown={index === data.length - 1} />
-      ))}
-    </>
-  );
-};
-
-export const TreeViewTab = () => {
+export const LuminaTreeViewTab = () => {
   const {
     state: { builderDataContext },
     dispatch
@@ -87,13 +17,13 @@ export const TreeViewTab = () => {
       <div className={styles.treeHead}>
         <h3 className={styles.treeTitle}>Components</h3>
         <span className={styles.treeAddButton}>
-          <AddComponentButton id={""} text="Add" />
+          <LuminaAddComponentButton id={""} text="Add" />
         </span>
       </div>
       <LuminaSearchBar />
-      <TreeviewHeader />
+      <LuminaTreeviewHeader />
       <div className={styles.treeScroll}>
-        <ComponentTree
+        <LuminaComponentTree
           // Confirmar se a data é undefined ou não
           data={
             builderDataContext.builderData[builderDataContext.selectedPage].children!
