+++ apps/lumina-core/src/staticComponentsPath/index.ts
@@ -9,7 +9,7 @@ import { config as configGrid } from "@/components/grid/config"
 import { config as configImage } from "@/components/image/config"
 import { config as configTitleText} from "@/components/titleText/config"
 import { config as configFlex } from "@/components/flex/config"
-import { TConfig } from "@/models/showEditModel"
+import { TConfig } from "@/lumina-models/lumina-showEditModel"
